+++ __tests__/e2e.ts
@@ -1,3 +1,5 @@
+import { toRiver } from '../src';
+
 describe('e2e', () => {
   it('todo', () => {
     expect(true).toBe(true);